Validating ScriptableObjects
Note: there is a bug that causes the CustomAttributes Checker to show a number corresponding to the amount of errors in ScriptableObject CustomAttributes Checker in the Artifice Validator window. The number should be part of the ScriptableObject CustomAttributes Checker.

The way that things are organized in this patchset is bad, for example how I use Artifice_ValidatorModule_CustomAttributeChecker.GenerateValidatorLogs(property, _validatorDrawerMap, Logs); or Artifice_ValidatorModule_CustomAttributeChecker.GenerateValidatorDrawerMap(); in Artifice_ValidatorModule_ScriptableObject.