Added some more SerializedPropertyExtensions (#142)
* Added some more SerializedPropertyExtensions

Added some SerializedPropertyExtensions that are more flexible which should help developers avoid repeating themselves more easily.

* Removed unnecessary cast from object to object

* Started using GetParentTarget in another place

* Made non-generic GetTarget method public (#138)

* Added a GetValue method for SerializedProperty

It felt strange that there is a way to get a value string but not to get the actual value (and it's something I need).

I also tried to reduce code duplication by making sure that this method is used in places where applicable.

* Update Artifice_SerializedPropertyExtensions.cs

* Toolbar of validator is now also supported for Unity 6000.3 and higher with new Unity's API (#139)

* Added fix for template class inheritance for enable if (#141)

* Added a GetValue method for SerializedProperty

It felt strange that there is a way to get a value string but not to get the actual value (and it's something I need).

I also tried to reduce code duplication by making sure that this method is used in places where applicable.

* Update Artifice_SerializedPropertyExtensions.cs

* Reverted most changes

I reverted most changes and simply made the uncast GetTarget method public. The untyped GetTarget method was made public since you don't always know the type you want/the property has at compile time.

Using GetTarget for GetValueString would result in less code duplication but it would also inevitably change some of the return values and I'm not sure if it's for the better.

---------

Co-authored-by: Zacharias Pervolarakis <39521184+ZackPer@users.noreply.github.com>

* Added some more SerializedPropertyExtensions

Added some SerializedPropertyExtensions that are more flexible which should help developers avoid repeating themselves more easily.

* Removed unnecessary cast from object to object

* Started using GetParentTarget in another place

* Update Artifice_CustomAttributeDrawer_OnValueChangedAttribute.cs

---------

Co-authored-by: Zacharias Pervolarakis <39521184+ZackPer@users.noreply.github.com>